+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Updated `traceRay` with additional parameter to define maximum localization error.
 - Divided `generateImage` functions in `createImage` and `saveImage`.
 - Added `renderBlack` to skip ray tracing when no intersections are found during GDS estimation.
+- `GDALALLRegister` is now ensured to be called only once per program.
 - Minor code polishing.
 
 ## v0.4.4